Move count_votes into its own module.
